# docs/swagger.yaml
definitions:
  components.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/components.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/validations.ValidationError'
        type: array
    type: object
  components.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 442902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  dto.CreateVehicleRequest:
    properties:
      location:
        type: string
      model:
        type: string
      name:
        type: string
      price:
        type: integer
      year:
        type: integer
    type: object
  dto.DBVehicleDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      location:
        type: string
      model:
        type: string
      name:
        type: string
      price:
        type: integer
      updated_at:
        type: string
      year:
        type: integer
    type: object
  dto.MemberAuth:
    properties:
      email:
        example: user@comecord.com
        type: string
      password:
        example: calista78Batista
        type: string
    type: object
  dto.MemberRegistration:
    properties:
      email:
        example: user@comecord.com
        minLength: 6
        type: string
      password:
        example: calista78Batista
        minLength: 6
        type: string
      phone:
        example: +7 (999) 999-99-99
        type: string
    required:
    - password
    type: object
  dto.Role:
    properties:
      name:
        example: Admin
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
  dto.RoleList:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
  dto.TotpRequest:
    properties:
      accountName:
        example: user@comecord.com
        type: string
      issuer:
        example: comecord.com
        type: string
    type: object
  validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /api/v1/health/:
    get:
      consumes:
      - application/json
      description: Health Check
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
      summary: Health Check
      tags:
      - Health Check
  /api/v1/members/login:
    post:
      consumes:
      - application/json
      description: Login a member
      parameters:
      - description: member
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.MemberAuth'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
      summary: Login a member
      tags:
      - Members
  /api/v1/members/register:
    post:
      consumes:
      - application/json
      description: Registration a member
      parameters:
      - description: member
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.MemberRegistration'
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
      summary: Registration a member
      tags:
      - Members
  /api/v1/members/totp/active/{code}:
    get:
      consumes:
      - application/json
      description: Active TOTP authentication for member
      parameters:
      - description: Code
        in: path
        name: code
        required: true
        type: string
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Active TOTP authentication
      tags:
      - Auth
  /api/v1/members/totp/generate:
    post:
      consumes:
      - application/json
      description: Generate TOTP authentication for member
      parameters:
      - description: payload
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.TotpRequest'
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
      summary: Generate TOTP
      tags:
      - Auth
  /api/v1/roles/create:
    post:
      consumes:
      - application/json
      description: Создание роли
      parameters:
      - description: role
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.Role'
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
      summary: Создание роли
      tags:
      - Roles
  /api/v1/roles/list:
    get:
      consumes:
      - application/json
      description: Вывод всех ролей
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/dto.RoleList'
            type: array
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
      summary: Вывод всех ролей
      tags:
      - Roles
  /api/v1/vehicles/:
    post:
      consumes:
      - application/json
      description: Create a vehicle
      parameters:
      - description: Create a vehicle model
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateVehicleRequest'
      responses:
        "201":
          description: Created response
          schema:
            allOf:
            - $ref: '#/definitions/components.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/dto.DBVehicleDTO'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/components.BaseHttpResponse'
      security:
      - BearerAuth: []
      summary: Create a vehicle
      tags:
      - Vehicles
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"